// App/src/app/shared/models/user.model.ts
import {Establishments} from './establish.model';
import {Review} from './review.model';
import {Booking} from './booking.model';
export class User {
        id: number;
        name: string;
        email: string;
        phoneNumber: string;
        password: string;
        isBlock: boolean;
        role: string;
        Establishment: Establishments[];
        reviews: Review[];
        bookingList: Booking[];
        constructor(name: string) {
                this.name = name;
        }
}